pyproject.toml
[tool.poetry]
name = "goit-ds-hw-03"
version = "0.1.0"
description = ""
authors = ["Fedichkin Denis <fedichkin.dy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pymongo = "^4.7.3"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

